adding new server socket and analysis -- preparing for test installation
